+++ docs/docs/CHANGELOG.md
@@ -2,4 +2,12 @@
 
 ### Feat
 
+- add a sort-of-working pycg
+- add the code2flow svg output
+- initialize with radon
 - initialize with copier
+
+### Refactor
+
+- report markdown and filter by min_score
+- run prdc post-install